Fixed most issues with scheduling.c
Along with fixing the issues, made adjustments to queue.c, queue.h, scheduling.c, scheduling.h, and commandline_parser.c. Added global.h which will hold global variables that can be used in multiple .c files, but was mainly made to allow the job queue to be global to the scheduling module and soon-to-be dispatching module.