Changed a lot
Added things to the makefile so it can now compile, debug, and run the C files that are necessary so far. Running "make" then "make run" in the terminal will compile the c files and then run the main. Running "make debug" in the terminal will compile the c files and then run the main in debug mode so when an error occurs it will show you where and you can then enter certain commands that I don't know yet to make it tell you more information. Also changed the help menu to print the commands that do and don't work as of now. Added bases cases to the list command and all scheduling policy commands to print comments if those bases cases are met, else it will run the "loop" part of the code in those functions. Changed the run command to malloc a new node instead of what was there before to fix the "bus error: 10" that would occur when trying to add a second job to the queue. The list command "works" without crashing but is printing unexpected results. will need to fix that before seeing if the scheduling policies are working so we can actually understand what is what.